+++ CHANGELOG.md
@@ -2,12 +2,13 @@
                   
 ## [1.0.1]
 
+- Fixed get active project timeout exception.
 - Missing `:` on configuration panel label.
 
 ## [1.0.0]
 
 ### Changed
-                                
+                                               
 - Fixed unsafe project source folders updates.
 - Moved settings file **mt4ij.xml** to **.idea** project folder (Thank you to **Vojtěch Krása** and **Petr Makhnev**).